# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexum REST API",
    "description": "REST API for the Nexum AI-powered development orchestration system. Provides CRUD operations for plans and tasks, status transitions, execution state inspection, agent claiming, and configuration access. All endpoints are versioned under `/api/v1/`. Write endpoints (POST, PUT, PATCH, DELETE) require Bearer token authentication via API key when authentication is enabled. Use GET /auth/status to discover authentication requirements.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "paths": {
...
    "/plans": {
      "get": {
        "operationId": "listPlans",
        "summary": "List all plans",
        "description": "Returns all plans, optionally filtered by branch and/or status.",
        "tags": ["Plans"],
        "parameters": [
          {
            "$ref": "#/components/parameters/QueryBranch"
          },
          {
            "$ref": "#/components/parameters/QueryPlanStatus"
          }
        ],
        "responses": {
          "200": {
            "description": "List of plans",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_Plan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "createPlan",
        "summary": "Create a new plan",
        "description": "Creates a new plan with the given name, branch, and goal. Generates a unique plan ID automatically. Returns 201 Created on success.",
        "tags": ["Plans"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanRequest"
              }
            }
          }
        },
        "security": [{"bearerAuth": []}],
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/plans/{branch}/{plan_id}": {
      "get": {
        "operationId": "getPlan",
        "summary": "Get a specific plan",
        "description": "Returns the details of a plan identified by branch and plan ID.",
        "tags": ["Plans"],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathBranch"
          },
          {
            "$ref": "#/components/parameters/PathPlanId"
          }
        ],
        "responses": {
          "200": {
            "description": "Plan details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch": {
        "operationId": "updatePlan",
        "summary": "Update a plan",
        "description": "Partially updates a plan. Only fields provided in the request body are changed.",
        "tags": ["Plans"],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathBranch"
          },
          {
            "$ref": "#/components/parameters/PathPlanId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanRequest"
              }
            }
          }
        },
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Updated plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deletePlan",
        "summary": "Delete a plan",
        "description": "Deletes a plan and all associated data (specs directory and state directory). Returns 204 No Content on success.",
        "tags": ["Plans"],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathBranch"
          },
          {
            "$ref": "#/components/parameters/PathPlanId"
          }
        ],
        "security": [{"bearerAuth": []}],
        "responses": {
          "204": {
            "description": "Plan deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/plans/{branch}/{plan_id}/status": {
      "patch": {
        "operationId": "transitionPlanStatus",
        "summary": "Transition plan status",
        "description": "Transitions a plan to a new status. The transition must be valid according to the plan status machine. Returns 422 if the transition is not permitted.",
        "tags": ["Plans"],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathBranch"
          },
          {
            "$ref": "#/components/parameters/PathPlanId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionPlanStatusRequest"
              }
            }
          }
        },
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Updated plan with new status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "PathBranch": {
        "name": "branch",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Git branch name. Must be non-empty, contain no '..' or leading/trailing '/', and only use alphanumeric characters, hyphens, underscores, or forward slashes."
      },
      "PathPlanId": {
        "name": "plan_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Unique plan identifier (e.g., 'PLAN-001')."
      },
...
      "QueryBranch": {
        "name": "branch",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Filter plans by git branch name."
      },
      "QueryPlanStatus": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": ["draft", "queued", "planning", "reviewing", "approved", "complete", "rejected"]
        },
        "description": "Filter plans by status."
      },
...
    },
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": ["error", "message", "status"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Short error type name (e.g., 'NotFound', 'Validation', 'Conflict').",
            "enum": ["NotFound", "BadRequest", "Conflict", "Internal", "Unauthorized", "MethodNotAllowed", "Validation"]
          },
          "message": {
            "type": "string",
            "description": "Full human-readable error message."
          },
          "status": {
            "type": "integer",
            "format": "uint16",
            "description": "HTTP status code."
          }
        }
      },
...
      "PlanResponse": {
        "type": "object",
        "required": ["id", "name", "status", "created", "branch", "goal", "scope", "background", "tasks"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique plan identifier (e.g., 'PLAN-001')."
          },
          "name": {
            "type": "string",
            "description": "Human-readable plan name."
          },
          "status": {
            "type": "string",
            "description": "Current plan status.",
            "enum": ["draft", "queued", "planning", "reviewing", "approved", "complete", "rejected"]
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 timestamp when the plan was created."
          },
          "branch": {
            "type": "string",
            "description": "Git branch associated with this plan."
          },
          "goal": {
            "type": "string",
            "description": "High-level goal the plan aims to achieve."
          },
          "scope": {
            "type": "string",
            "description": "Description of the scope and boundaries of the work."
          },
          "background": {
            "type": "string",
            "description": "Background context explaining why this plan exists."
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskReferenceResponse"
            },
            "description": "Ordered list of task references belonging to this plan."
          }
        }
      },
      "TaskReferenceResponse": {
        "type": "object",
        "required": ["id", "name", "completed"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique task identifier."
          },
          "name": {
            "type": "string",
            "description": "Human-readable task name."
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the task has been completed."
          }
        }
      },
      "CreatePlanRequest": {
        "type": "object",
        "required": ["name", "branch", "goal"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the plan. Must be non-empty."
          },
          "branch": {
            "type": "string",
            "description": "Git branch to associate with the plan. Must pass branch name validation."
          },
          "goal": {
            "type": "string",
            "description": "High-level goal the plan aims to achieve."
          },
          "scope": {
            "type": "string",
            "description": "Optional description of the scope and boundaries."
          },
          "background": {
            "type": "string",
            "description": "Optional background context."
          }
        }
      },
      "UpdatePlanRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New name for the plan."
          },
          "goal": {
            "type": "string",
            "description": "New goal statement."
          },
          "scope": {
            "type": "string",
            "description": "New scope description."
          },
          "background": {
            "type": "string",
            "description": "New background context."
          }
        },
        "description": "Only fields provided will be updated."
      },
      "TransitionPlanStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "description": "Target plan status.",
            "enum": ["draft", "queued", "planning", "reviewing", "approved", "complete", "rejected"]
          },
          "by": {
            "type": "string",
            "description": "Optional identifier of the actor performing the transition."
          }
        }
      },
...
      "ListResponse_Plan": {
        "type": "object",
        "required": ["items", "total"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "description": "Total number of items."
          }
        }
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "error": "NotFound",
              "message": "Not found: plan not found",
              "status": 404
            }
          }
        }
      },
      "BadRequest": {
        "description": "Malformed or invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "error": "BadRequest",
              "message": "Invalid request: malformed JSON",
              "status": 400
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "error": "Internal",
              "message": "Internal error: IO operation failed",
              "status": 500
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request failed schema validation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "error": "Validation",
              "message": "Validation error: Invalid plan status transition: 'draft' -> 'complete'",
              "status": 422
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "error": "Unauthorized",
              "message": "Missing Authorization header. API key required.",
              "status": 401
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Include the API key in the Authorization header: 'Authorization: Bearer <api-key>'. Authentication is enabled via config and protects write endpoints by default."
      }
    },
...
  }
}